refactor: remove google recaptcha
